requirements.txt
numpy==1.23.4
pandas==1.5.2
Pillow==9.4.0
torch==1.13.1
